Organise teleport scripts
