+++ game/src/main/kotlin/world/gregs/voidps/world/activity/transport/teleport/SkillsNecklace.kts
@@ -1,4 +1,4 @@
-package world.gregs.voidps.world.activity.transport
+package world.gregs.voidps.world.activity.transport.teleport
 
 import world.gregs.voidps.engine.client.message
 import world.gregs.voidps.engine.data.definition.AreaDefinitions
